Added excercise on virtual destructors
https://www.quantstart.com/articles/C-Virtual-Destructors-How-to-Avoid-Memory-Leaks
From QuantStart Team

Inheritace also lends itself to virtual methods, where implementation is provided by any specific sublasses.
However, once an inheritance hierarchy is created, with memory allocation occuring at each stage in the
hierarchy, it is necessary to be very careful about how objects are destroyed so that any memory leaks are avoided.
In order to achieve this, we make use of virtual destructorl